wip: Employee entity, REST controller parameter "merchType"

+++ src/OzonEdu.MerchandiseService.Domain/Models/Entity.cs
@@ -7,7 +7,7 @@ namespace OzonEdu.MerchandiseService.Domain.Models
     public abstract class Entity
     {
         int? _requestedHashCode;
-        public virtual int Id { get; protected set; }
+        public virtual long Id { get; protected set; }
 
         private List<INotification> _domainEvents;
         public IReadOnlyCollection<INotification> DomainEvents => _domainEvents?.AsReadOnly();
